requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pypdf
sqlalchemy
chromadb
sentence-transformers
tiktoken
google-generativeai>=0.5.0
google-auth>=2.0.0
